update gis client & configuration
